+++ src/ext/java/org/opentripplanner/ext/transmodelapi/model/DefaultRoutingRequestType.java
@@ -363,7 +363,7 @@ public class DefaultRoutingRequestType {
                         .newFieldDefinition()
                         .name("debugItineraryFilter")
                         .type(Scalars.GraphQLBoolean)
-                        .dataFetcher(env -> request.debugItineraryFilter)
+                        .dataFetcher(env -> request.itineraryFilters.debug)
                         .build())
                 /*
                 .field(GraphQLFieldDefinition.newFieldDefinition()
@@ -400,12 +400,6 @@ public class DefaultRoutingRequestType {
                                 "Whether to apply the ellipsoid->geoid offset to all elevations in the response.")
                         .type(Scalars.GraphQLBoolean)
                         .dataFetcher(env -> request.geoidElevation)
-                        .build())
-                .field(GraphQLFieldDefinition.newFieldDefinition()
-                        .name("transitGeneralizedCostLimit")
-                        .description("A relative maximum limit for the generalized cost for transit itineraries. The limit is a linear function of the minimum generalized-cost.")
-                        .type(Scalars.GraphQLString)
-                        .dataFetcher(env -> request.transitGeneralizedCostLimit)
                         .build())
                 /*
                 .field(GraphQLFieldDefinition.newFieldDefinition()