+++ src/main/java/org/opentripplanner/routing/algorithm/mapping/RoutingRequestToFilterChainMapper.java
@@ -1,11 +1,9 @@
 package org.opentripplanner.routing.algorithm.mapping;
 
 import org.opentripplanner.model.plan.Itinerary;
-import org.opentripplanner.routing.algorithm.filterchain.AdjustedCost;
 import org.opentripplanner.routing.algorithm.filterchain.GroupBySimilarity;
 import org.opentripplanner.routing.algorithm.filterchain.ItineraryFilter;
 import org.opentripplanner.routing.algorithm.filterchain.ItineraryFilterChainBuilder;
-import org.opentripplanner.routing.api.request.GroupBySimilarityParams;
 import org.opentripplanner.routing.api.request.RoutingRequest;
 
 import java.time.Instant;
@@ -26,20 +24,20 @@ public class RoutingRequestToFilterChainMapper {
       Consumer<Itinerary> maxLimitReachedSubscriber
   ) {
     var builder = new ItineraryFilterChainBuilder(request.arriveBy);
+    var p = request.itineraryFilters;
 
     // Group by similar legs filter
-    if(request.groupBySimilarity != null) {
-      GroupBySimilarityParams p = request.groupBySimilarity;
+    if(request.itineraryFilters != null) {
 
-      AdjustedCost adjustedCost = AdjustedCost.create(p.minSafeTransferTimeFactor);
+      builder.withMinSafeTransferTimeFactor(p.minSafeTransferTimeFactor);
 
-      if (p.keepOne >= 0.5) {
+      if (p.groupSimilarityKeepOne >= 0.5) {
         builder.addGroupBySimilarity(
-            new GroupBySimilarity(p.keepOne, KEEP_ONE, adjustedCost)
+            new GroupBySimilarity(p.groupSimilarityKeepOne, KEEP_ONE)
         );
       }
 
-      if (p.keepNumOfItineraries >= 0.5) {
+      if (p.groupSimilarityKeepNumOfItineraries >= 0.5) {
         int minLimit = request.numItineraries;
 
         if (minLimit < 0 || minLimit > MIN_NUMBER_OF_ITINERARIES) {
@@ -47,18 +45,19 @@ public class RoutingRequestToFilterChainMapper {
         }
 
         builder.addGroupBySimilarity(
-            new GroupBySimilarity(p.keepNumOfItineraries, minLimit, adjustedCost)
+            new GroupBySimilarity(p.groupSimilarityKeepNumOfItineraries, minLimit)
         );
       }
     }
 
     builder
         .withMaxNumberOfItineraries(Math.min(request.numItineraries, MAX_NUMBER_OF_ITINERARIES))
-        .withTransitGeneralizedCostLimit(request.transitGeneralizedCostLimit)
+        .withMinSafeTransferTimeFactor(p.minSafeTransferTimeFactor)
+        .withTransitGeneralizedCostLimit(p.transitGeneralizedCostLimit)
         .withRemoveTransitWithHigherCostThanBestOnStreetOnly(true)
         .withLatestDepartureTimeLimit(filterOnLatestDepartureTime)
         .withMaxLimitReachedSubscriber(maxLimitReachedSubscriber)
-        .withDebugEnabled(request.debugItineraryFilter);
+        .withDebugEnabled(p.debug);
 
     return builder.build();
   }